Refactor event classes to be in model
Moving the event classes to be under the model component ensures consistency in the architecture.
